Get characteristic/cfScript to build given latest HOL
Issue was use of "do" as a variable name, and monadic-syntax is now on
by default, reserving do as a delimiter keyword. It's easy enough to
turn monadsyntax off, so this is the fix at the head of the script
file.

If people are often wanting to use this variable name, it's possible
that this change should be done once and for all in preamble.sml