Next version of CF for non-terminating programs
Currently, the only thing of interest in the final limit state st' of
the Div (diverge) case is the io_events. This focus on only one
component in the state is (I think) unavoidable because there is no
limit state for the oracle state or reference store.

It is clear that this use of the ffi state requires the io_events list
to be an llist, but there might be problems with the current
definition of the IO (those problems can probably be solved).