Yet another CF setup for non-terminating programs
The previous version had the problem that a "limit state" had to fit
into the state type of the source semantics. A quick experiment where
types where the type of io_events in ffiTheory was changed to io_event
llist showed that representing limit states in the source would be
very messy (e.g. involving changing all compiler IL semantics
functions) and lead to some ugly cases in the source semantics.

This latest attempt (this commit) rephrases the CF setup so that limit
states are only represented in the heap type of CF. This approach
makes the changes local to CF: only the CF setup needs to be able to
represent "limit states".

Context: Johannes and I are trying to setup CF so that diverging
programs are verified w.r.t. a postcondition that talks about a state
that comes at the end of an infinite number of states in the
non-terminating program. In these commit messages, I call these
imaginary final states "limit states".