Fix last uncheck constructor cheat.
I'm not convinced that the semantics for equality is right. In
particular, it can return type error when two constructors of different
types are compared. I don't immediately see how this can get compiled
into a simple tag check, so if there is a problem in later phases, we
will have to change it.