Make xlet_auto less likely to generate extra assumptions
find_equality_types somewhat heuristically searches for EqualityType
assumptions and tries to prove them, as part of xlet_auto machinery (for
discharging the first subgoal out of xlet once the postcondition has
been invented).

In stackProg it was generating a spurious
  EqualityType ($= : v -> v-> bool)
assumption, which breaks the HOL4's new TAC_PROOF that checks for extra
assumptions.