Prove a correctness result for a cmlPtreeConversion function
The result states that the conversion function that takes a parse tree
to an AST type value has a right-inverse (a printing function if you
will), as long as the AST value is "user-expressible". Some AST values
in this type can't be expressed in the parse trees. For example, the
built-in type constructors TC_int etc are outside of the range of the
conversion, as is the "for internal use" Tvar_db.

Moreover, show that the ptree_Type function always hits
user-expressible types.

In retrspect, we should have proved analogous theorems for all of the
conversion functions as part of our correctness argument for the
parser.