Prove that new heap after GC has no dead data (algo level proof)
This commit strengthens old results which say that the
number of reachable elements in the old heap is the same as
the total number of heap elements in the heap after GC.

The new result states that the *length* (i.e. number of words)
of the heap after GC is the same as the length of the reachable
part of the heap before GC.

This is a vital property for when we want to prove that if
something fits next to all the reachable elements then it
definitely will fit after a full GC.

At present, this has only been proved for the algorithm level.
The property needs to be transported down to the implementation
level to be used for the purpose mentioned above.