Finish refactor of evaluate proof in clos_callProof
I found where it matters that env_rel is for the free vars in the
compiled expression (ys) instead of the source expression (xs). There
is a cheat there now for that stuck case. The problem is that if I
change it to ys, then the optimised case of App where the pure check
succeeds is not provable. Maybe the goal statement can be tweaked
again to get App to work in pure case with ys for env_rel.