Complete proofs in cmlPtreeConversionProps
Also remove use of heap in semantics/proofs because this depends on
theories in semantics that don't build at the moment.  One also has to
use

    Holmake --no_prereqs

in semantics/proofs to stop it from complaining about failures in
semantics.

Finally: slightly refactor grammar as signature declarations inside
structures conform to the same grammar as those that appear at
top-level.