Address state_rel problem with code_includes
This commit removes the problematic code_includes from state_rel and
instead pushes the current code table through v_rel and friends. This
way, if there is an optimised closure value, then its corresponding
code is part of the current code table. This approach gets around the
problem, but makes the proofs a bit more ugly.

There are several cheats in this commit. Most of them are routine
cases (If, Handle, Raise) that have yet to be updated for the new
v_rel. There are also some simple theorems about v_rel that are
cheated. These simple cheats are marked with comments "routine proof".