Make basis and ffi-divergence play nice together
...at least up to and including TextIOProof.

It's nice to see that, with some minor tweak of xlet_auto
and friends, the proofs in basis require only very minor changes
to account for the addition of FFI divergence --- as should be the case,
since nothing in basis FFI-diverges. Mostly boils down to an extra
xsimpl here and there to discharge trivial ==f> side-conditions
generated by xlet_auto.

Some of the specs of higher-order functions could be generalised to
account for FFI-divergence. For example, the spec of map f xs could
say that it FFI-diverges if some for some x in xs, f(x) FFI-diverges.
However, this is likely to cause nothing but pain for the 99% of use
cases where map does not thus diverge; more thinking needed to determine
when to bother making specs that account for divergence in basis.