Start reworking clos_callProof for better compile_inc
Previously the incrementral compiler had as its state the same g as
the initial compiler, which carries around all of the generated code
in g. Using such a state for the incrementral compiler is terrible as
all of the already installed code would need to be kept forever.

My original plan was to split subg into two parts: one for the num_set
and one for the generated code. Howver, that turned out to require
essentially redoing the entire clos_call proof.

I opted for a more local approach whereby the incrementral compiler
has only the num_set in its state and, in the proof, recreates the
code part from the current code. This approach seems to work, but has
the slightly awkward case of specifying the code part for the limit
state.