Make whole_prog_thm automation support ffi-diverging programs
The automation checks whether the conclusion of your goal is a
whole_prog_spec or a whole_prog_ffidiv_spec and then uses
slightly different rules in a few steps depending on that.

This mechanism assumes that your CF spec is either a POSTf or a
POSTv -- there is currently no support for things written with POST
that can do either.

However, a simple workaround for this is to prove two specs:
one for POSTf and one for POSTv (under appropriately different
assumptions), call whole_prog_thm separately with the two
corresponding whole_prog_spec thms. Then (if desired) a semantics
theorem about all executions can be constructed from those two
by case analysis.