Remove Install from closSem and patSem
The hope is to temporarily disable Install so that we can get a
cheat-free version of the compiler proofs (where the install-related
assumptions become trivial). I have attempted to preserve the previous
state cleanly in comments.

This commit does not finish clos_to_bvlProof: some more work will be
required in hacking theorem statements to remove assumptions to be able
to do that.