Another stab at the semantics
The capture-avoiding type instantiation function turns out to be useless,
since INST is already capture-avoiding.

This variant does not type-instantiate terms before intepreting them in the fragment.
Rather, choose a type instantiation at top-level and then lazily applies it
to all non-composite terms it encounters as the term semantics is constructed.

This trick allows variables with the same name but different types to be
given different semantics, even when interpreted under a type instantiation that
makes them one and the same. This appears to be necessary in order to prove the
Abs rule sound -- otherwise I run into a situation where the same interpretation must
be given to unrelated free occurrences of a variable name (with different types)
in the hypotheses and the conclusion.