Work on ag32_ffi_write_code_thm
It really needs to be done with some semantics/simulation, because of
writes to memory. So now doing it using the semantics theorems for the
subroutines. I think there is necessarily some overlap/duplication
between the pure shallow proofs and the shallow<->deep connection,
basically because code is in memory.