Fix backend/proofs for concurrent THENs
All of this directory builds with multithreading on, and THENs running
in parallel.

Fixes in
- bvi_letProof (rpt (impl_tac >- ...))
- bvl_handleProof (TRY (conj_tac >- ...))
- clos_ticksProof (TRY (... >- ...))
- clos_to_bvlProof (TRY (... >- ...))
- data_to_word_bignumProof ( interesting!  See below )
- flat_exh_matchProof (TRY (... >- ...))
- flat_to_patProof (srw_tac[][Once rwt,...])
- pat_to_closProof (first_x_assum (fn th => ... >- ...))
- stack_removeProof (rpt (... >- ...))
- word_allocProof (TRY (impl_tac >- ...))
- word_removeProof (TRY (.. >- ...))
- word_simpProof (TRY (... >- ...))
- word_to_stackProof (TRY (... >- ...))

In data_to_word_bignumProof, there is a fix to workaround a
combination of problems with the cache (very stateful!) in the
arithmetic decision procedure and the strange behaviour of
full_simp_tac when it causes a case-split on something in the
assumptions that turns out to be a disjunction.  The workaround is to
call the full_simp twice to ensure that the arithmetic d.p. fires with
or without the cache having been primed.

See github issue HOL-Theorem-Prover/HOL#606 for the full_simp_tac
problem.