Add read/write modes to fsFFI
Now the filesystem tracks the mode (read or write) for open fds and only
allows reading/writing if the mode is correct. The hope is that this
allows us to implement ffi_read for ag32 which only/always reads stdin
(and ignores the requested fd) even on filesystems where stdout/stderr
are also present.

This commit updates basis/ but not examples/ or other uses of the basis.