Add --types command-line arg to compiler
This --types option causes the bootstrapped compiler to stop after
type inference and just print the types of all the top-level bindings
in the program. I hope this makes the compiler easier to use.

The printed type information contains all of the bindings from the
basis program plus the bindings for the user-supplied program.