Fix a few proofs in lab_to_targetProof
Closes #536

This commit bring #536 to a close. The inlining of constants
has been done for clos_known, but was in the end not done for
lab_to_target because there the initial compiler is shared
with the incremental compiler to such an extent that inlining
the mandatory initial values for the initial compiler just
caused a mess.