Test and tweak new EqualityType prover.
Fix the first interesting bug: the slightly unusual type sizes
defined in basicSizeTheory break things, which can be fixed up.

Also adjust the way that the proof is wrapped, so that failure
to construct a proof strategy is caught/ignored (unsupported types)
but failure to complete the tactic proof is exposed as an error.