Fix/refactor simple evaluate proofs for 'local'.
The additional recursion creates new work in terminationScript,
evaluatePropsScript. These proofs are now fixed.

I didn't like these proofs, and changed a lot of them. The evaluate
and evaluate_decs functions are composed in an almost-monadic way
out of elements that treat the clock in a particular way. The new
proof defines a property for this, proves that it is compositional,
and thus reduces the number of inductive proofs over evaluate_decs
substantially. Unfortunately there's some higher-order stuff floating
around, which eventually leads to manual applications of BETA_RULE,
which I'm not a huge fan of.