Move wordLang-related syntax lemmas to data_to_wordProof
This is only the first step because many of the lemmas at the
end of data_to_wordProof ought to be moved to the different
word_*Proof theories.

There was one complicated case: assign_get_code_label is a
function over the op datatype from closLang. I didn't know
where to put it so I put it in closLangTheory, because it is
a function directly related to the syntax only. An alternative
would have been to put it in closProps, but that would have
caused bvl, bvi and dataProps to open closProps, which we do
not want for such a simple syntacctic reason.