Rewrite tail-recursive functions in terms of tailrec
The point is that it's easier to prove div specs about functions where
it's easy to isolate the effects of a single loop iterations. Functions
on tailrec form satisfy this.

We only care about tail-recursive functions because if your function ain't
tail-recursive, it will not actually diverge but run out of stack.