Restore ML_code_env constant.
This was previously folded into the ML_code predicate, but doing so
requires computing environment merges when new blocks are opened,
which in hindsight might be expensive. Let's revert to the ML_code_env
constant and the use of merge_env of all open environments, which will
then be examined one at a time on each nsLookup.