Add a how-to file to root of CakeML repo
The idea is that this file is a concise description of what one
needs to know to start using the bootstrapped CakeL compiler as
a normal compiler. The text assumes the reader knows some SML or
OCaml. I think this text should be part of the tar.gz files that
contains the bootstrapped compiler.

The text provides several code examples that will be compiled
and run as part of regression tests in order to prevent this
text from becoming out-of-date.

Closes #433