Make List.map and List.mapi evaluate left-to-right
I was confused by the fact that

    List.map print ["line 1\n", "line 2\n", "line 3\n"];

printed:

    line 3
    line 2
    line 1

This was of course due to right-to-left evaluation in the definition of
`List.map` which by default has implementation: `f x :: map f xs`.

This confuses inserts a `let` to force `f x` to be evalauted first.

Now extract_code.cml can use `List.map print ...`.