Fix grepProg by retranslating balanced_map
The retransltion is not an ideal solution. Ideally, the relevant
functions in the grep example should use the mlmap interface to
the map datatype to avoid the retranslation. However, this is
complicated because mlmapTheory lives in CakeML while large
parts of the grep example live in HOL. Maybe mlmapTheory should
be moved to HOL.