Discontinue iocat (no longer compatible with TextIO)
The problem with the iocat example is that it requires access
to internals of TextIO. If the TextIO module is to have a good
design then we cannot expose the internal buffer nor primitives
such as `read` and `write`.

It will remain in git history as an interesting example that
worked at one point in time.