Defines a bootstraped version of Skew Binomial Heaps
The existing implementation of priority queues (Skew Binomial Heaps)
supports insertion in O(1) and merge, find_min, delete_min in O(log n).
We define a new implementation that uses our previous implementation
internally to support insert, merge and find_min in O(1) and delete_min
in O(log n).

The bootstraped heap is either empty or it contains an element and a
skew binomial heap of bootstraped heaps.