Defines a function to get a bag from a bootstrap binomial heap
In order to prove the functional correctness of bootstraped
binomial heaps, we need a function to transform them into
bags.

To prove the termination of this function, I introduced another
function that computes the size of a bootstraped binomial heap.
This function counts the number of bootstraped heaps (empty or
not) that occur inside a bootstraped heap.

I had to verify some properties on this size function to prove
the termination.