Change the order required on the elements
This commit cleans loosen the constraints the comparison operator that compares the elements in the priority queue. This was necessary because when we will have to prove the correctness of the delete operation, we will need the fact that an ordered boostraped heap stores its children in an ordered skew binomial heap.
To prove that merge and insert preserve this ordering, we needed to prove that b_comparison_heap (a function that creates a comparator between bootstraped heaps from a comparison between elements) maintains the weak linearity of the relation.
This was not possible so I switched to a total pre-order, which turns out to be enough.