Prove some lemmas about the relation between heaps and bootstraped heaps
Since we use our first implementation of heaps in our bootstraped version,
we need to make a link between the definition of the minimal element of the
underlying heap and the minimal element of the bootstraped implementation
that uses it.
In order to prove the lemmas expressing these links, we hade to strenghten
the invariant on our boostraped heaps (in the is_b_heap_ordered property)
and to extend the proofs of ordering for links and merge accordingly.