Prove that deleting on bootstraped heaps preserves ordering
This commit contains a lot of auxilliary lemmas regarding the conservation of ordering on bootstraped heaps when we apply functions of our first implementation on the underlying skew binomial heap.
It also defines and proves the relation between the result of delete_min and find_min on bootstraped heaps, which was needed to prove that delete_min preserves the ordering of a boostraped heap.