Extirpate intermediate theorem syntax
This commit replaces all(?) occurrences of the old-style Theorem syntax

Theorem name
 `thm_statement
 `
 (tactic)

with

Theorem name:
 thm_statement
Proof
 tactic
QED

Instances of the Theorem syntax that get piped into forward reasoning
are treated specially, as per a suggestion by @mn2000. So

Theorem name
  `thm_statement`
  (tactic)
  |> forward_rules;

maps to

Theorem name = Q.prove(`
  thm_statement`,
  tactic)
  |> forward_rules;

This will produce silly output in some cases (if the forward
rules end in a call to save_thm, the theorem gets saved twice),
but it seems not worth it to try to detect this automatically.