New approach to calls_correct proof
The proof hypothesises a 'future point' at which the analysis
is completed. This is complicated by future installs. There was
a plan to use the clock to pick a future point after all installs,
but instead I just adjusted the proof to let the 'future point'
walk forward through the induction.