Modify code of b_inputUntil and some bIO heapconditions
This commit changes b_inputUntil to be implemented in a similar way
to inputLine, for ease of proof. It also changes IOFS_WO_iobuff
to IOFS in the pre/post conditions of b_refillBuffer_with_read_spec and
b_input1_spec. Additionally, b_input1_spec is renamed to b_input1_IOFS_spec,
and a new spec b_input1_spec is introduced, which is similar to b_input_IOFS_spec,
but with STDIO instead of IOFS in pre and postcondition.