Define `map_diff` and prove its termination
In order to prove the termination of the version of Dijkstra that doesn't use
the decreaseKey operation on priority queues, we need to define the difference
operation on balanced maps.
We had to prove a few lemmas about the size of balanced maps when operations are
applied on them to prove the termination of this diff operation.