Start defining the compilation strategy for #667
I started defining the skeleton of the compilation strategy described in (Maranget, 2008) but I added the let expressions generation in the decision trees (which is not present in the paper).
To do that in a smart (let's say not dumb) way, we need to gather informations on the different branches, and how they bind names to constructs. That is the purpose of the column_infos type, which inspect the first column, and does three things :

- collect all the names refering to a "default" pattern, a variable or a as-clause over a variable
- the list of all constructors in the column
- collect all the names refering to a constructor pattern for each of the constructors in the column

In this example :

```
match l1 with
  C1(p1, p2, p3) as x =>
  C1(p1, p2, p4) as y =>
  (C2(p1', p2', p3') as z) as z' =>
  w =>
  v as u =>
  _ as t =>

```

we should obtain a informations about this info about the column :

- "default names" : `[w, v, u, t]`
- constructors and their associated names : `[(C1, [x, y]), (C2, [z, z'])]`

I also took or-patterns into account (I think, nothing is verified yet).

Next step is using these informations to correctly apply the compilation
strategy recursively for each of the constructors and for the default case.