Build bootstrap translation with string type in AST
Curiously, this removes the explicit explode that issue #353
is about. However, don't be fooled: the explode now exists
inside the translator generated code because the translator
inserts it in order to patch up the difference between
LIST_TYPE CHAR (char list type in CakeML) and
HOL_STRING_TYPE (string type in CakeML).

This commit makes progress on #666.