Parametrize the size function by a constructor arity
In the size function, we need to take into account that a wildcard pattern
can be expanded to a list of other wildcard patterns in the spec operation.
To take that into account, we parametrize the size function by an arity. We
adapt the proofs for the equality of size when columsn are swapped and the
decreasing size when default op is applied