Start implementation and proof of flat_to_clos phase
This compiler phase essentially only does one thing: it switches from
explicit variables names of flatLang to dB indexing of closLang. It
assumes that no pattern matching appears in the flatLang code.

The compiler proof is straighforward and uses a new way of structuring
the proof: each case of the main induction on evaluate is split into
a separate theorem. Such a structure should make the proof nicer to
develop and maintain.