Refactor Icing rewriter into a separate state component
Previously the Icing components ("fp_*") where part of the global state.
While this worked out for the proofs so far up to implementing a source to
source Icing pass, it a) cluttered up the proofs and b) broke the translator.

This commit refactors the Icing components into a separate parameter of type "fp_state"
separate from the global state in "evaluate".