Prove Equal_code_lemma with stack usage bound
The bound is based on the clock value given to word_eq. By word_eq_thm this
bounds the stack usage of Equal to something like

  MustTerminateLimite * stack_frame_size(Equal)

which isn't the most practical bound ever.

The idea is that we can obtain useful bounds by proving a version of word_eq_thm with
tighter bounds phrased in terms of the structure of a dataLang value.