Patch up broken proof in iocatProg (see comments)
The real issue here is that various rewrites and other things done
internally in cfLetAutoLib seems to depend on _when_ cfLetAutoLib
was loaded. When the Lib is loaded appears to have changed after a
recent update Holmake. This makes the xlet_auto heuristics instantiate
some variables in some theorems with the wrong things.