Fix translator proofs that were broken by adding fp value trees
The key lemma that was necessary to be fixed was `evaluate_empty_state_IMP`
which required a separate `intro` lemma for running in a floating-point
environment, with optimizations enabled, but not performing any.
The definition of `WORD` further does not fit floating-point values anymore
thus this commit introduces a `DOUBLE` translator type.
However it currently cannot properly be handled and might have to be changed in
the future.