Cheat-free assign_RefArray
This commit moves the AllocVar outside the Call and into the
body of the callee for the op. This circumvents some cutset-related
pain and allows size_of to make better approximations of the actual
heap usage.

The previous formulation of RefArray_thm is now relatively useless,
since RefArray does much less work. Sadly I was not able to figure
out a new formulation that made sense, so I ended up moving all the
reasoning into the (now huge) assign_RefArray proof instead.
Hopefully this doesn't make it too unmaintainable.