Explicitly construct miniBasis with ml_progLib
This theory is comprised of copypasta from the corresponding
ProgScripts in basis/.

The reason this exists is to allow using CF to reason about programs
on top of the mini-basis.  CF proofs are not in general portable
across environments, so any CF proofs done in the prog state
corresponding to the "real" basis cannot be easily ported to the
mini-basis that's actually being used for cyes.