Rework semantics and translator to have better equality for doubles
Previously floating-point equality was defined structurally for the value trees.
This made compilation non-semantics preserving.
The fix is to make do_eq compare only compressed values in the source semantics.
In combination with making `NoOpt` scopes apply eager compression, this allowed
to go back to the old translator implementation of floating-points as 64 bit
words. The floating-point examples in ml_translator_test now are working too.