Lots of edits, do_eval/end_eval is working
So, it looks like the invariants are basically right so that
adding an eval generation with do_eval and exiting it with
reset_env_generation, and saving an env with Denv all work
out ok.

The main missing thing is that source Eval also saves the
final env and returns it, which will require some stub code
to make it work.