Start reworking reader to use buffered I/O
The reader now always reads all input, but tokenizes it line by line.
Includes functions which does this on regular non-buffered I/O, to
support stdin (and running on ag32).

The proofs are a bit chaotic in this commit, and some theorems are
cheated. There's still a lot of unused definitions lying around from the
first version of the reader.