Retire EnvLookup
This was a good idea for use in a REPL, but conflicts with how the
compiler actually works. A mapping from names to values simply isn't
how environments are implemented, and the only way to coerce values
back 'up' from the lower/later compiler levels is to pass code 'down'
to them, i.e. Eval, which already exists.

The solution for a REPL probably involves adjusting the code being
run to save and print values as they are computed.