Require source to call compiler in Eval mode
This is an admission that attaching the compiler text to a program
text, and sharing the AST datatypes, etc, creates too many headaches.

Instead, we require the source program to include the compiler,
probably by building on its translation. The source must provide
the compiler input/output pair to Eval. This is clumsy, but
can be solved by existing tools for constructing programs.