Begin changing oracles to partial functions
This means an oracle can include only finitely many valid Eval/Install
steps and then forbid further steps. This means we do not have to
invent further oracle steps in the case where the program would only
call Eval a limited number of times, and in particular we do not have
to establish safety of these invented steps.