Add a total compiler in source_eval
Add a hook where an additional compiler can be provided and
used to invent "dummy" data to extend an oracle to infinity.
Unlike the standard compiler, this additional compiler must be
total. Prove that the generated oracle is well-formed in that
every step is taken by one of the two compilers.