Restore do-install next-state assertion
Every level of do-install asserts that the next-state in the
oracle equals the one that would be produced by the compiler.
This was considered a bit awkward because of partiality, but
since we're making things total anyway, and the assertion
spans all the levels of the compiler, the assertion needs
to be added up to the source semantics.