PIC-compatible x64 cml_main
movabs cannot be used in position-independent code without text
relocations as it embeds an absolute address into the text segment.
This was originally added to support heaps+stacks greater than 2GiB, but
since 01f9fa8e the heap and stack are no longer in the data segment and
the data segment size is negligible.

Fixes #438.