Add Char.some to basis
This commit adds the following function to the Char module:

    Char.some: char -> char option

The Char.some function returns its char argument wrapped in a Some
constructor. It does so *without* performing on-the-fly allocation;
instead it looks up the returned value from a pre-computed vector.

The implementation is written in a way that makes this code easily
dead-code eliminated if it is unused. Thus the pre-computed vector
should only be present in memory if it is actually used.

This function is meant to be used in TextIO.