Working backendProofScript
With a bit of specialisation of the early proof phases, everything
fits together for a working proof of compile_correct.

Note that this implicitly assumes that Eval is disabled, since the
prim_sem_env setup leaves Eval set to NONE. The source_eval proofs
aren't actually used, and the oracle used is a nonsense value (see
"TODO syntax oracle") since it is never actually needed.