Move no-eval proof up, use name_ind_cases
If an execution succeeds with eval_state set to NONE, it will
succeed identically with any other eval state. This was proven
previously in source_eval and is needed by translator proofs,
so it is moved up to evaluateProps. This was also a good time
to make use of name_ind_cases now in HOL4 to delete explicit
appearances of Case/elim_Case in previous work.