Implement foldChars, foldLines, foldTokens for TextIO
These new functions allow for simple processing of stdin
(last argument is `None`) or a file's content (last argument
is `Some filename`). The result is None in case the file
could not be opened. Any exception that is raised by the
supplied functions will be propagated out (after consuming
the rest of the input and closing of the file, if a file
is bring read).

    TextIO.foldChars: (char -> 'a -> 'a) -> 'a -> string option -> 'a option
    TextIO.foldLines: (string -> 'a -> 'a) -> 'a -> string option -> 'a option
    TextIO.foldTokens: (char -> bool) -> (string -> 'a) -> ('a list -> 'b -> 'b) -> 'b -> string option -> 'b option

Under the hood, these are implemented in terms of buffered
I/O for speed.