Rewrite expression ptree conversion
Applying functions that consume lists to the parse tree nodes apparently
sends the definition machinery into meltdown. It doesn't prove
termination on its own, and it generates completely mad induction
theorems that in turn sends the translator into a meltdown. I changed
parts of the PEG to avoid 'rpt', and in turn avoid some list-functions
in ptree_Expr.