Change identifier rules to those of HOL Light
HOL Light changes the OCaml frontend to allow mixed-case variable names
(mostly to support ALL_CAPS_NAMES) and allow anything for identifiers
that is not a valid type constructor/module name. The latter are those
identifiers that start with a capital letter and then contain no
uppercase letters (but perhaps lowercase letters, digits, underscores
and ticks).