Add HOL Light-specific tokens to lexer.
A few identifiers are to be treated as infixes: THEN, ORELSE, etc.
The parser does not yet pick up these tokens. We'll also have to
add some logic to the parser to read parenthesized infix tokens --
e.g. '(' 'THEN' ')' -- as if they were an value-name "THEN". (Just as
you would with infix symbols.)