Add HOL Light infixes to the PEG
They're not yet in parse tree conversion, because I have yet to finish
how to deal with infix operators being put into prefix position: i.e.
by putting parens around them. Any operator in prefix position can be
used as a value name, so it must be possible to assign to it, etc.