translator: define type-rep functions
In addition to the type-representation invariants (e.g. LIST_TYPE),
the translator now also defines type-representation functions which
map to the v type, for any type for which it could previously prove
EqualityType.

This is needed in the Eval proofs to represent the config oracle as
a sequence of v values, to avoid further type parameters.