Delete Constant op from closLang
The original plan was to introduce large constants as closLang's
Constant op, but that plan changed once it became apparent that we
want to treat closed lambdas (Fn) as constants too. So the new plan is
to have clos_annotate discover the large constants by annotating the
existing ops. This means that the Constant op is no longer needed.

Progress on issue #803