Tidy up candle_prover_semanticsScript.sml
There are some existing definitions (the basis decls, the candle decls,
etc.) that can be used. The main theorem is still cheated in the
diverging case, and there is a cheat about the candle kernel environment being
env_ok. The state_ok proof is cheated until the translator is set up to
reveal the contents of simple references from the basis (which should all be
ok).