Install pp_X pretty-printers early in translation
Sets up the pp_data type and the initial pretty printers and
translates them immediately after decProg. Some of these can
be replaced later, as is already done for Vector.

Requires translating some of the toString functionality in
mlint earlier in the process. Some of those functions were
adjusted to be easier to translate.

Working up to IntProg, which is somehow broken by some of its
constants already being translated, or for some other mysterious
reason, shadowing induction theorems doesn't seem to be doing
what it should (or rather, what it previously did).