Add no-fix as an option in fixes
The fixes mapping is now option-typed, with NONE indicating
that a type is not printable and that no simple fix for it
exists. The pretty-printer then prints these with a sensible
default message (<unprintable val of type ...>) which seems
like a reasonable default (e.g. printing TextIO.instream).

The current version generally seems robust enough.