Handle wildcards in special-case patterns
Unwrap 'Comb _' to 'Comb _ _' (and the same for other, special CakeML
2-argument constructors).

All types with arity >1 are made tupled by the OCaml parser. We have
some special cases for types defined in CakeML, e.g. the constructors
of the Candle kernel types, and Append of app_list. For example, before,
we would unwrap 'Comb (x,y)' to 'Comb x y', but 'Comb _' would fail.