Add AallocFixed to source AST and semantics
This is the same as Ref in closLang and below, i.e. a call to
AallocFixed with arguments v1, v2, ..., vn will create an
array that is of length n and contains v1, v2, ..., vn.

This new op is not typeable at present.

It is added to the CakeML soucre so that the PureCake compiler
can reliable target it. An alternative would have been to make
the compiler smarter and turn some uses of Aalloc intoclosLang's
Ref op, which seems a bit brittle considering that the PureCake
compiler really must hit Ref in closLang to produce decent code.