Prevent the pattern parser from taking apart alias
Writing something like '(x, (y as z))' would cause the pattern parser
to unwrap the 'x as z' pattern a second time, giving back something
similar to '(x, y) as z'. This change makes the parser finalize alias
patterns in the same way as it finalizes product patterns, so that they
are not opened up by grabPairs a second time.