Improvement (again) in printing failure handling
There are situations where the injected print code for "val _ = _ : _" may fail
to type check, e.g. if the necessary pp_X functions have somehow been shadowed
and got the wrong type, or maybe other ways.

To (yet again) try to handle this better, type check each val-print declaration
in isolation, handling the type error by replacing just that declaration with
a safer version from a range of 3 progressively safer options.

This should provide a little more information to the user in this case, rather
than a single error and the loss of information for the whole block.