[Pancake] Completed the lexer definition and proofs.
We have a top-level function pancake_lex for turning a string into
a token stream.

Some trickiness with locations was resolved; record syntax should
not be used since it does not reduce arithmetic expressions on
columns or rows.

I have added additional tokens which became necessary as I
implemented the to-be-committed parser.