Complete source-to-source oracle proofs
A better approach is to have source_evalProof export a helper proof
that establishes that semantics_prog is preserved under an oracle
change if that change adds a compiler which preserves evaluate_decs.
This then requires no new induction in source_letProof and
source_to_soureProof, just an application of that result.

This closes the cheats on this branch, I think.