Change how the caml parser deals with guards
Previously, encountering a row with a guard would duplicate the match
expression corresponding to all rows below the guarded row. In a
scenario where all rows were guarded, the expression would grow
exponentially. With this change, the parser binds the remainder of the
match expression to a lambda, so that growth becomes linear (hopefully).