Move compute expressions + semantics into own dir
This division looks a bit awkward (compute_syntax/ and compute/) but the
idea is this: the actual compute expressions and their interpreter code
do not depend on Candle functions, and are to be translated /before/
Candle. The rest of the compute code (under compute/) does depend on
Candle, however, and must be translated after the Candle functions.

With this change, the type, term and theorem types will be translated
before the usual place in the Candle ml_hol_kernel_funsProgScript.sml