Use loads/needs instead of #load/#need in REPL
This makes the REPL pick up on the words 'needs' and 'loads' when
they sit at the top level. If they are followed by a string literal
and a double semicolon token, the REPL loads a file (like previously
with #load and #need).

The motivation behind this change is to make sure Candle accepts HOL
Light source files with as minor changes as possible.